==> inst/comparison/2-run-magic.ipynb <==
import magic
import pandas as pd
import matplotlib.pyplot as plt
import time
start_time = time.time()
X = pd.read_csv("~/rerun/zeisel_raw.csv",header = 0,index_col=0)
X = X.transpose()
magic_operator = magic.MAGIC()
X_magic = magic_operator.fit_transform(X)
print("--- %s seconds ---" % (time.time() - start_time))

pd.DataFrame.to_csv(X_magic.transpose(), "~/rerun/zeisel_magic.csv")
